BOARD AND GAME
Make the Board class independent by not loading the player in the change_cell method

Create initialize, play, next_to_play and not_a_tie methods to switch between players and also validate if the game is a tie in the Game class

optimize  detect winner  and modify running methods to test the Game class.

The code passes all rubocops.